requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-cors-headers==3.5.0
django-dotenv==1.4.2
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
flex==6.14.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonpointer==2.0
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.7
packaging==20.4
pandas==1.2.3
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rfc3987==1.3.8
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.2
validate-email==1.3
wincertstore==0.2
xlrd==2.0.1
